Internal Documentation

Documentation for LAP_julia.jl's internal interface.

Index

LAP_julia.Flow

LAP_julia.Image

LAP_julia.angle_mean

LAP_julia.angle_rms

LAP_julia.clean_using_gaussain

LAP_julia.data_gen.gen_chess

LAP_julia.data_gen.gen_rand_flow

LAP_julia.gradient_points.find_keypoints_from_gradients

LAP_julia.gradient_points.gradient_magnitude

LAP_julia.inpaint.inpaint_nans!

LAP_julia.interpolation.warp_img

LAP_julia.lap.multi_mat_div

LAP_julia.lap.polyfilter_lap

LAP_julia.lap.single_lap

LAP_julia.lap.window_sum!

LAP_julia.mean

LAP_julia.mse

LAP_julia.pad_images

LAP_julia.rescale_intensities

LAP_julia.vec_len

LAP_julia.visualise.imgshow

LAP_julia.visualise.imgshowflow

LAP_julia.visualise.showflow

LAP_julia.visualise.warp_imgshowflow

Functions

LAP_julia.gradient_points.gradient_magnitude — Function

gradient_magnitude(f)

Return the directional derivatives (components of the gradient) and the gradient mag for both color and grayscale images. For spatialorder(f)=="xy", grad[1] is a derivative wrt x, the first coordinate and grad[2] wrt to the second...

LAP_julia.pad_images — Function

pad_images(image1::Image, image2::Image)

Adds zeros to the right and bottom of image1 and image2 to make them the same size.

LAP_julia.rescale_intensities — Function

rescale_intensities(image1::Image, image2::Image)

Rescale image1 and image2 intensities to span the whole [0, 1].

LAP_julia.clean_using_gaussain — Function

clean_using_gaussain(u::Matrix{<:Number}, window_half_size_one_dim::Integer)

Clean the Matrix u by smoothing using a square 2D Gaussian filter of size 2 * window_half_size_one_dim + 1 in each dimension.

clean_using_gaussain(u::Matrix{<:Number}, window_half_size)

Clean the Matrix u by smoothing using a 2D Gaussian filter of size 2 * window_half_size + 1.

LAP_julia.mse — Function

function mse(x, y)

Calculates the mean squared error between x and y.

LAP_julia.angle_rms — Function

function angle_rms(x, y)

Calculate the root mean square error in angle between x and y. Output in degrees.

LAP_julia.angle_mean — Function

function angle_mean(x, y)

Calculate the mean error in angle between x and y. Output in degrees.

LAP_julia.vec_len — Function

function vec_len(x)

Calculate the lenght of vector x. x is a complex number.

LAP_julia.mean — Function

function mean(x)

Calculate the mean of x.

LAP_julia.inpaint.inpaint_nans! — Function

inpaint_nans!(flow::Flow)

Inpaint NaNs in flow using surrounding non-NaN values.

References

Oliveira, Manuel & Bowen, Brian & McKenna, Richard & Chang, Yu-Sung. (2001). Fast Digital Image Inpainting.. 261-266. Here.

See also: showflow, Flow

LAP_julia.interpolation.warp_img — Function

function warp_img(img, dx, dy; border_strat::Symbol=:replicate)

Warp the image img by dx in x direction and by dy in y direction.

border_strat indicates how to act when the resulting coordinate outside of bounds of img. Values: :replicate, :zeros.

See also: showflow, imgshowflow, Flow

LAP_julia.lap.multi_mat_div — Function

multi_mat_div(A, b)

Return E, where E[i, :] is the solutution of the least squares problem $\min\|D_(i)x - c_(i)\|^2$ for each $D_(i)$ and $c_(i)$, where $D_(i)$ is A[:, :, i], $c_(i)$ is b[:, i] and $i$ is the size of the second dimension of b and third dimension of A. In other words, each row of E is the solution to one matrix from A and it's corresponding vector from b.

LAP_julia.lap.window_sum! — Function

function window_sum!(filter_result, pixels, image_size, window_size)

Filter the array pixels of shape image_size with a filter that performs a sum of values of the area window_size with each pixel as the center. Store the result in filter_result. The filtering uses "symmetric" padding.