Internal Documentation

Documentation for LAP_julia.jl's internal interface.

Functions

LAP_julia.angle_mae — Method

function angle_mae(x, y)

Calculate the mean absolute error in angle between x and y. Output in degrees.

LAP_julia.angle_rmse — Method

function angle_rmse(x, y)

Calculate the root mean square error in angle between x and y. Output in degrees.

LAP_julia.back_substitution3d — Method

back_substitution3d(A, b)

Solve the linear systems of equations $C_n x_n = d_n$, where the matrixes $C$ (in row echelon form) are slices from the first 2 dimensions of A, the vectors $d$ are the slices from the first dimension of b and $n$ is the size of the third dimension of A. Returns the coefficients $x_n$ as a 2D matrix coeffs, where the vectors $x_n$ are the rows of this matrix.

See also: gem3d!, multi_mat_div2

LAP_julia.compare_dicts — Method

compare_dicts(old_dict, new_dict)

Compare two dicts by dividing their corresponding values, return the comparation dict.

Note: (old / new)

LAP_julia.cumsum2d! — Method

cumsum2d!(result, A)

Cummulative sum over 2D matrix A, storing the result into result.

LAP_julia.fill_circle! — Method

fill_circle!(mag, pos, spacing, marker)

Fill a circle with a radius of spacing around the point pos into the matrix mag with marker.

LAP_julia.fun_on_dict_values — Method

fun_on_dict_values(dicts, fun)

Run function fun on all same keys of all dictionaries dicts.

using BenchmarkTools
fun_on_dict_values((Dict(:a => 1), Dict(:a => 12)), mean)

# output

Dict{Any,Any} with 1 entry:
  :a => 6.5

LAP_julia.gem3d! — Method

gem3d!(A, b)

Reduce the linear systems of equations $C_n x_n = d_n$ to the row echelon form. The matrixes $C$ are slices from the first 2 dimensions of A, the vectors $d$ are the slices from the first dimension of b and $n$ is the size of the third dimension of A.

See also: gem3d!, back_substitution3d

LAP_julia.gradient_magnitude — Method

gradient_magnitude(f)

Return the directional derivatives (components of the gradient) and the gradient mag for both color and grayscale images. For spatialorder(f)=="xy", grad[1] is a derivative wrt x, the first coordinate and grad[2] wrt to the second...

LAP_julia.inds_to_points — Method

inds_to_points(inds::Array{CartesianIndex, 1})

Transform an array of CartersianIndexes to an array of where each column is a vector of indices of the input array.

LAP_julia.inpaint_nans! — Method

inpaint_nans!(flow::Flow)

Inpaint NaNs in flow using surrounding non-NaN values.

References

Oliveira, Manuel & Bowen, Brian & McKenna, Richard & Chang, Yu-Sung. (2001). Fast Digital Image Inpainting.. 261-266. Here.

See also: showflow, Flow

LAP_julia.interpolate_flow_rbf — Method

interpolate_flow_rbf(flow_at_inds,
                     inds::Array{CartesianIndex, 1},
                     flow_size,
                     rbf_method::T=Multiquadratic(2)) where {T <: ScatteredInterpolation.RadialBasisFunction}

Interpolate a complex displacement field of size flow_size from displacement vectors flow_at_inds at inds using the a Multiquadratic RBF model with the parameter ε and using only the values at inds for the interpolation.

Examples of method from ScatteredInterpolation:

Multiquadratic(ɛ = 1)

\[ϕ(r) = \sqrt{1 + (ɛr)^2}\]

Polyharmonic(k = 1)

\[ϕ(r) = r^k, k = 1, 3, 5, ... \\ ϕ(r) = r^k ln(r), k = 2, 4, 6, ...\]

(See ScatteredInterpolation for more.)

See also: showflow, Flow, interpolate_flow, interpolate_flow_quad

LAP_julia.max_displacement — Method

max_displacement(flow)

Find the maximum displacement of flow, ignoring NaNs.

LAP_julia.multi_mat_div2 — Method

multi_mat_div2(A, b)

Solve the linear systems of equations $C_n x_n = d_n$, where the matrixes $C$ are slices from the first 2 dimensions of A, the vectors $d$ are the slices from the first dimension of b and $n$ is the size of the third dimension of A. Returns the coefficients $x_n$ as a 2D matrix coeffs, where the vectors $x_n$ are the rows of this matrix.

Note: Uses Gaussian elimination and back substitution. Note: For small matrices $C$, this is an order of magnitude faster than

See also: gem3d!, back_substitution3d

LAP_julia.multi_mat_div_using_qr — Method

multi_mat_div(A, b)

Return E, where E[i, :] is the solutution of the least squares problem $\min\|D_ix - c_i\|^2$ for each $D_i$ and $c_i$, where $D_i$ is A[:, :, i], $c_i$ is b[:, i] and $i$ is the size of the second dimension of b and third dimension of A. In other words, each row of E is the solution to one matrix from A and it's corresponding vector from b.

LAP_julia.normalize_to_zero_one — Method

normalize_to_zero_one(arr)

Normalize arr so that it has a maximum of 1 and minimum of 0.

LAP_julia.pad_images — Method

pad_images(image1::Image, image2::Image)

Adds zeros to the right and bottom of image1 and image2 to make them the same size.

LAP_julia.prepare_gaussian_filters — Method

prepare_gaussian_filters(filter_half_size)

Get KernelFactors for the first 3 forward and backward gaussian filters from the LAP paper.

LAP_julia.rescale_intensities — Method

rescale_intensities(image1::Image, image2::Image)

Rescale image1 and image2 intensities to span the whole [0, 1].

LAP_julia.smooth_with_gaussian! — Method

smooth_with_gaussian!(A::Matrix{<:Number}, window_half_size_one_dim::Integer)

Clean the Matrix A by smoothing using a square 2D Gaussian filter of size 2 * window_half_size_one_dim + 1 in each dimension.

LAP_julia.smooth_with_gaussian! — Method

smooth_with_gaussian!(A::Matrix{<:Number}, window_half_size)

Smooth the Matrix A by smoothing using a 2D Gaussian filter of size 2 * window_half_size + 1.

LAP_julia.to_lin_index — Method

to_lin_index(cart_ind, size)

Get linear index from CartesianIndex cart_ind of 2D matrix of size size.

LAP_julia.vec_len — Method

function vec_len(x)

Calculate the lenght of vector x. x is a complex number.

LAP_julia.window_sum! — Method

function window_sum!(result, pixels, img_size, window)

Get a sum of values (using a "symmetric" padding on the borders) in a window window around each point and saving the sum into result.

Note: Uses a filter algorithm with a 2D ones kernel. Note: For small images (around 256x256) and small filters (window of size 9 and below) is faster than window_sum3!. <- factcheck this again

LAP_julia.window_sum3! — Method

function window_sum3!(result, pixels, img_size, window)

Get a sum of values (using a "symmetric" padding on the borders) in a window window around each point and saving the sum into result.

Note: Uses a cumsum algorithm. Note: For small images (around 256x256) and large filters (window of size 11 and above) is faster than window_sum!. <- factcheck this again

LAP_julia.window_sum3_at_inds! — Method

function window_sum3!(result, pixels, img_size, window, inds)

Get a sum of values (using a "symmetric" padding on the borders) in a window window around inds and saving the sum into result.

Note: Uses a cumsum algorithm.