Public Documentation

Documentation for LAP_julia.jl's public interface.

See the Internals section of the manual for internal package docs covering all submodules.

Index

LAP_julia.Flow

LAP_julia.Image

LAP_julia.data_gen.gen_chess

LAP_julia.data_gen.gen_rand_flow

LAP_julia.gradient_points.find_edge_points

LAP_julia.inpaint.inpaint_nans!

LAP_julia.interpolation.interpolate_flow

LAP_julia.interpolation.warp_img

LAP_julia.lap.polyfilter_lap

LAP_julia.lap.single_lap

LAP_julia.mean

LAP_julia.visualise.imgshow

LAP_julia.visualise.imgshowflow

LAP_julia.visualise.showflow

LAP_julia.visualise.warp_imgshowflow

Public Interface

Types

LAP_julia.Flow — Type

Flow{T} = Matrix{Complex{T}} where T <: Real

Flow is a Matrix with elements that are Complex.

LAP_julia.Image — Type

Image{T} = Matrix{T} where T <: Real

Image is a Matrix with elements that are Real.

Functions

LAP_julia.mean — Method

function mean(x)

Calculate the mean of x.

LAP_julia.lap.polyfilter_lap — Method

polyfilter_lap(target::Image, source::Image; filter_num::Integer=3, max_repeats::Integer=1, display::Bool=true)

Find a transformation flow (complex displacment field), that transforms image source to image target.

Arguments

target::Image: the image we want source to look like.

source::Image: warped image we want to transform into target.

filter_num::Integer=3: the number of basis filters used in single_lap calls (so far only =3 implemented).

max_repeats::Integer=1: the maximum number of times an iteration of one filter size can be repeated.

display::Bool=true: use verbose prints and return an array of figures.

Outputs

flow::Flow: is the complex vector field that transforms source closer to target.

source_reg::Image: is the image source transformed by flow.

[figs::Matrix{Figure}: is a 2D array of PyPlot Figures which shows the work of the algorithm at each iteration.

For each iteration there are 3 Figures in this order: 1) current u_est, 2) newest addition to u_est Δ_u, 3) current source_reg.]

Describtion

Implements the basic concept of Algorithm 2 from the paper without some features. It uses single_lap iteratively; in each iteration using the transformation estimated by single_lap to warp the source image closer to the target image and then using this warped closer image as the source image in the next iteration, while using progressively smaller filter_half_sizes to estimate even small and faster varying displacements.

See also: single_lap, imgshow,imgshowflow, warp_imgshowflow, Flow

LAP_julia.lap.single_lap — Function

single_lap(image_1::Image, image_2::Image, filter_half_size::Integer, window_size, filter_num::Integer=3)

Return an estimate of a smoothly varying flow of size of image_1 which is the displacement that transforms image_2 closer to image_1.

Arguments:

image_1::Image: grayscale image 1.

image_2::Image: grayscale image 2.

filter_num::Integer=3: the number of basis filters used (so far only =3 implemented).

filter_half_size: the half size of the base of the gaussian filters used.

window_size: the size of the local window (list of 2 ints) usually same as filter_size.

See also: polyfilter_lap, showflow, warp_imgshowflow, imgshowflow

LAP_julia.visualise.imgshow — Method

imgshow(img; <keyword arguments>)

Return a figure with image img. Plot origin is in the botom left.

Arguments

fig=nothing: add a figure to plot in. By defaults creates a blank new figure.

figtitle::String="Image": add title to the figure.

ret::Symbol=:figure: set return object, by default returns Figure, other options: :pyobject returns a PyObject. (Using figure makes Juno directly plot.)

See also: showflow, imgshowflow, warp_imgshowflow

LAP_julia.visualise.imgshowflow — Method

imgshowflow(img, flow; <keyword arguments>)

Return a figure with an image img and displacement field flow.

Arguments

img: the image to be plotted.

flow::Flow: the vector flow to be plotted.

skip_count=nothing: the number of vectors to skip between each displayed vector. By default set so that the output is $20 × 20$ vectors.

fig=nothing: add a figure to plot in. By defaults creates a blank new figure.

mag::Real=1: magnify the plotted vectors.

key::Bool=true: add key with maximum vector length.

figtitle::String="Iamge with Flow": add title to the figure.

ret::Symbol=:figure: set return object, by default returns Figure, other options: :pyobject returns a PyObject. (Using figure makes Juno directly plot.)

See also: imgshow, showflow, warp_imgshowflow

LAP_julia.visualise.showflow — Method

showflow(flow::Flow; <keyword arguments>)

Return a figure with the displacement field flow by default skipping some vectors to make it easy to read.

Arguments

flow::Flow: the vector flow to be plotted.

skip_count=nothing: the number of vectors to skip between each displayed vector. By default set so that the output is $20 × 20$ vectors.

fig=nothing: add a figure to plot in. By defaults creates a blank new figure.

mag::Real=1: magnify the plotted vectors.

key::Bool=true: add key with maximum vector length.

figtitle::String="Flow": add title to the figure.

ret::Symbol=:figure: set return object, by default returns Figure, other options: :pyobject returns a PyObject. (Using figure makes Juno directly plot.)

See also: imgshow, imgshowflow, warp_imgshowflow

LAP_julia.visualise.warp_imgshowflow — Method

warp_imgshowflow(img, flow; <keyword arguments>)

Return a figure with an image and a displacement field, where the image is warped by the displacement field.

Arguments

img: the image to be plotted.

flow::Flow: the vector flow to be plotted.

skip_count=nothing: the number of vectors to skip between each displayed vector. By default set so that the output is $20 × 20$ vectors.

fig=nothing: add a figure to plot in. By defaults creates a blank new figure.

mag::Real=1: magnify the plotted vectors.

key::Bool=true: add key with maximum vector length.

figtitle::String="Iamge with Flow": add title to the figure.

ret::Symbol=:figure: set return object, by default returns Figure, other options: :pyobject returns a PyObject. (Using figure makes Juno directly plot.)

See also: imgshow, showflow, imgshowflow

LAP_julia.inpaint.inpaint_nans! — Method

inpaint_nans!(flow::Flow)

Inpaint NaNs in flow using surrounding non-NaN values.

References

Oliveira, Manuel & Bowen, Brian & McKenna, Richard & Chang, Yu-Sung. (2001). Fast Digital Image Inpainting.. 261-266. Here.

See also: showflow, Flow

LAP_julia.gradient_points.find_edge_points — Method

find_edge_points(img::Image; spacing::Int=10, number::Int=100, sigma=1, mask::BitArray{2}=[])

Greadily locates number of points in img with the highest gradient magnitude. These points have will be atleast spacing pixels appart, they are returned as an array of CartesianIndex. A mask can be used to exclude some parts of img from the search.

Note: sigma is the gaussian filter used to smooth img before looking for gradients.

LAP_julia.data_gen.gen_chess — Function

gen_chess(tile_size::Integer=50, board_size::Integer=4)

Create a chessboard image with board_size tiles in each dimension, where each tile is tile_size pixels in each dimension. Note that board_size has to be even.

See also: imgshow, gen_rand_flow

LAP_julia.data_gen.gen_rand_flow — Method

gen_rand_flow(flow_size::Tuple{T, T}=(200, 200), max_magnitude::Real=20, tile_size=nothing; filter_amp=nothing)::Flow where {T <: Integer}

Generate a smoothly varying random flow. The flow parameters are set by the function arguments.

It works by generating a tiled flow where each tile has a random uniform flow. Then it smooths these tiles by filtering with a gaussian.

Arguments

flow_size::Tuple{T, T}=(200, 200): dimensions of the flow.

max_magnitude::Real=20: maximum allowed amplitude of the displacement.

tile_size=nothing: size of random uniform flow tiles that make up the start matrix. (The larger the slower the flow.) Note: If set to flow_size or larger it will generate a uniform pixel shift in a random direction.

filter_amp=nothing: size of the gaussian filter which is used to smooth the random start matrix.

See also: showflow, Flow, gen_chess

LAP_julia.interpolation.interpolate_flow — Method

interpolate_flow(flow::Flow, inds::Array{CartesianIndex, 1}, method::T=Multiquadratic(2)) where {T <: ScatteredInterpolation.RadialBasisFunction}

Interpolate a flow using the a Multiquadratic RBF with the parameter ε and using only the values at inds for the interpolation.

Examples of method from ScatteredInterpolation:

Multiquadratic(ɛ = 1)

\[ϕ(r) = \sqrt{1 + (ɛr)^2}\]

Polyharmonic(k = 1)

\[ϕ(r) = r^k, k = 1, 3, 5, ... \\ ϕ(r) = r^k ln(r), k = 2, 4, 6, ...\]

(See ScatteredInterpolation for more.)

See also: showflow, Flow

LAP_julia.interpolation.warp_img — Method

function warp_img(img, dx, dy; border_strat::Symbol=:replicate)

Warp the image img by dx in x direction and by dy in y direction.

border_strat indicates how to act when the resulting coordinate outside of bounds of img. Values: :replicate, :zeros.

See also: showflow, imgshowflow, Flow